Small refactoring and code style (#3)
