+++ src/app/templates/common.py
@@ -1,0 +1,15 @@
+def load_style() -> str:
+    return """<style>
+    .stButton>button {
+        color: #F4F6F7;
+        background-color: #8E44AD;
+        border-color: #8E44AD;
+        # width: 10em;
+        width: auto;
+        display: block;
+        margin: auto;
+    # .stButton button:hover {
+    #     background-color:#018749;
+    #     color: black;
+    # }
+    }</style>"""
